requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.12.1,<3.13
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.1
Pillow>=5.3.0,<8.1.3
flake8>=3.6.0,<3.9.9
djangorestframework-jwt>=1.11.0,<1.12
django-debug-toolbar>=3.2.1,<3.3
black